refactor(common): eliminate db manager duplication via abstract base
- Introduce AbstractDatabaseManager with all shared SQL operations:
  getPlayer, getPlayerByName, isPlayerDead, arePlayersDead, revivePlayer,
  setLives, setFirstJoin, setLastSeen, setGraceUntil, getDeadPlayers,
  getPluginVersion, invalidateDeathStatusCache
- MySQLManager retains only MySQL-specific init, upsert syntax, and
  createMetadataTableIfNeeded DDL
- SQLiteManager retains only SQLite-specific init, upsert syntax, DDL,
  and overrides getBatchSize() to enforce 900 IN-param limit
- Fix malformed DeathStatusCache (duplicate imports, broken inner class)
- Resolves SonarCloud duplicated lines on new code
